ajuste delteById

+++ back-end-customer/src/main/java/com/cliente/api/service/ClienteService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -35,11 +36,16 @@ public class ClienteService {
         return obj.orElseThrow(() -> new ObjectNotFoundException("Não existe Cliente na base de dados com este Email"));
     }
 
-    public void delete (Long id) {
-        repository.deleteById(id);
+    public void delete (Long id) throws Exception {
+        Optional<Cliente> obj = Optional.ofNullable(findById(id));
+
+        if (obj.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new ObjectNotFoundException("Não eixste Clietne na base de Dados com este ID.");
+        }
     }
-
-
+    
     public Cliente update(Cliente cliente) throws Exception {
         findById(cliente.getId());
